+++ api/src/main/java/org/btrplace/model/constraint/MinMigrations.java
@@ -19,7 +19,11 @@
 package org.btrplace.model.constraint;
 
 /**
- * An objective that just minimizes the number of migrations to perform inside a reconfiguration.
+ * An objective that just minimizes the cumulative duration of the migrations
+ * to perform during a reconfiguration.
+ *
+ * Contrarily to {@link MinMTTR}, it only focuses on migration and do not consider
+ * the possible action delay.
  *
  * @author Fabien Hermenier
  */